Extract shared backup format helpers into backup-format.ts
packFiles, unpackFiles, and buildBackupBundle are now shared between
backup.ts and auto-backup.ts.
